chore: move components
