+++ packages/editor-ui/src/components/VariablesForm.vue
