+++ packages/editor-ui/src/components/VariablesUsageBadge.vue
